+++ admin/partials/device-detector-admin-settings-core.php
@@ -20,7 +20,7 @@
 					'action' => 'do-save',
 					'tab'    => 'core',
 				],
-				admin_url( 'options-general.php' )
+				admin_url( 'admin.php' )
 			)
 		);
 		?>