+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to **Device Detector** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **Device Detector** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased - will be 1.1.2]
+## [Unreleased - will be 1.2.0]
+### Added
+- Compatibility with PerfOps.One suite.
+### Changed
+- New menus (in the left admin bar) for accessing features: "PerfOps Analytics", "PerfOps Tools" and "PerfOps Settings".
 ### Removed
 - Compatibility with WordPress versions prior to 5.2.